requirements-cloud.txt
streamlit>=1.28.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
numpy>=1.21.0,<1.25.0
Pillow>=9.0.0,<10.0.0
protobuf>=3.20.0,<4.0.0
attrs>=21.4.0